Spotless, added a command to set winch speed, added button to smartdashboard

+++ src/main/java/frc/robot/subsystems/climber/ClimberBase.java
@@ -38,6 +38,5 @@ public abstract class ClimberBase extends SubsystemBase {
 
   public abstract void stopRightWinch();
 
-  public abstract void adjustWinch(int adjust);
-
+  public abstract void adjustWinch(double adjust);
 }
